requirements.txt
requests==2.28.1
azure-functions==1.11.0
azure-cosmos==4.0.0